model-lineage/pyproject.toml
[project]
name = "model-lineage"
version = "0.1.0"
description = "Model lineage scraping pipeline for HuggingFace models"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "huggingface-hub>=0.20.0",
    "neo4j>=5.15.0",
    "dvc>=3.40.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "tqdm>=4.66.0",
    "pyyaml>=6.0.0",
    "beautifulsoup4>=4.12.0",
]
